добавлены уведомления, добавлен класс Timberman
